+++ lib/model/annotation_task.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'label_image.dart';
+import 'labeled_image.dart';
 part 'annotation_task.freezed.dart';
 
 @freezed
