+++ lib/model/labeled_image.dart
