+++ test/main/stateholder/main_event_handler_test.dart
@@ -5,7 +5,7 @@ import 'package:ikut_annotation_v2/main/stateholder/main_event_handler.dart';
 import 'package:ikut_annotation_v2/main/stateholder/main_ui_model.dart';
 import 'package:ikut_annotation_v2/main/stateholder/main_ui_model_state_notifier_provider.dart';
 import 'package:ikut_annotation_v2/model/annotation_task.dart';
-import 'package:ikut_annotation_v2/model/label_image.dart';
+import 'package:ikut_annotation_v2/model/labeled_image.dart';
 import 'package:ikut_annotation_v2/model/my_error.dart';
 import 'package:ikut_annotation_v2/model/my_exception.dart';
 import 'package:mocktail/mocktail.dart';
